Update README.md
The Friday following Thanksgiving is referred to as "Black Friday" in American slang. 
It traditionally signals the beginning of the holiday shopping season in the US. Many 
retailers open early, often as early as midnight or even on Thanksgiving, and frequently
have heavily advertised bargains with steep discounts. Data from customers who purchased products
on this day might be examined to provide a quick statement of their preferences for specific products. 
Here, we have examined the data tuples of customers together with their purchasing criteria and quantity.This data is examined and forecasted solely to offer clients special discounts on goods based on their preferences and purchasing power. 
Data analysis currently makes use of data science with machine learning (ML), which teaches the computer using models and offers better
accuracy than business intelligence