README.md
The majority of brain strokes are caused on by an 
unanticipated obstruction of the heart's and brain's regular 
operations. researchers can use a variety of machine learning 
techniques to forecast the likelihood of a stroke occurring. This 
report studies the use of machine learning techniques to predict 
the long-term outcomes of stroke victims. We can accurately 
predict stroke, which is important for early treatment, by 
developing and analyzing several machine learning models. The
dataset included 5110 observations of patients who had suffered a 
stroke and their modifiable risk factors.